fixed by Yuval's review

+++ src/main/java/frc/robot/turret/TurretConstants.java
@@ -9,12 +9,13 @@ public class TurretConstants {
     static final double MIN_DEGREE = -135;
     static final int TOLERANCE_DEGREE = 3;
     static final int FLIP_POINT = 360;
+
     public static final class TurretComponentsA{
         static final int PID_IDX = 0;
-        static final int KP = 0;
-        static final int KI = 0;
-        static final int KD = 0;
-        static final int KF = 0;
+        static final double KP = 0;
+        static final double KI = 0;
+        static final double KD = 0;
+        static final double KF = 0;
         static final int ACCELERATION = 0;
         static final int CRUISE_VELOCITY = 0;
         static final int ACCELERATION_SMOOTHING = 0;
